Modularise library and correct vec_mul
